requirements.txt
accelerate==1.13.0
acres==0.5.0
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
ci-info==0.4.0
click==8.3.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
faiss-cpu==1.13.2
filelock==3.25.2
fitz==0.0.1.dev2
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.18.0
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.1.1
protobuf==6.33.5
prov==2.1.1
psutil==7.2.2
puremagic==2.1.0
pyarrow==23.0.1
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.27.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyxnat==1.6.4
PyYAML==6.0.3
rank-bm25==0.2.2
rdflib==7.6.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.5
tqdm==4.67.3
traits==7.1.0
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
